Patch 0.65

+++ src/main/java/lab/space/vilki_palki/service/impl/BannerServiceImpl.java
@@ -2,6 +2,10 @@ package lab.space.vilki_palki.service.impl;
 
 import jakarta.persistence.EntityNotFoundException;
 import lab.space.vilki_palki.entity.Banner;
+import lab.space.vilki_palki.mapper.BannerMapper;
+import lab.space.vilki_palki.model.banner.BannerResponse;
+import lab.space.vilki_palki.model.banner.BannerSaveRequest;
+import lab.space.vilki_palki.model.banner.BannerUpdateRequest;
 import lab.space.vilki_palki.repository.BannerRepository;
 import lab.space.vilki_palki.service.BannerService;
 import lab.space.vilki_palki.util.FileUtil;
@@ -9,20 +13,26 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.UUID;
+
+import static java.util.Objects.nonNull;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class BannerServiceImpl implements BannerService {
     private final BannerRepository bannerRepository;
+    private final BannerMapper bannerMapper;
 
     @Override
-    public List<Banner> getAllBannersByOrderByCreateAt() {
+    public List<BannerResponse> getAllBannersByOrderByCreateAt() {
         log.info("---------------Get All Banners Order By createAt---------------");
-        return bannerRepository.findAll(Sort.by(Sort.Direction.DESC, "createAt"));
+        return bannerRepository.findAll(Sort.by(Sort.Direction.DESC, "createAt"))
+                .stream()
+                .map(bannerMapper::toDto)
+                .toList();
     }
 
     @Override
@@ -33,25 +43,40 @@ public class BannerServiceImpl implements BannerService {
     }
 
     @Override
-    public void saveBanner(String name, MultipartFile image) {
+    public BannerResponse getBannerDto(Long id) {
+        log.info("---------------Get Banner Dto By Id" + id + "---------------");
+        return bannerMapper.toDto(getBannerById(id));
+    }
+
+    @Override
+    public void saveBanner(BannerSaveRequest request) {
         log.info("---------------Save Banner---------------");
-        Banner banner = new Banner();
-        banner.setName(name);
-//        if (FileUtil.saveFile(image.getOriginalFilename(), image))
-            banner.setImage(image.getOriginalFilename());
+        Banner banner = new Banner().setName(request.name());
+
+        if (nonNull(request.image())
+                && nonNull(request.image().getOriginalFilename())
+                && !request.image().getOriginalFilename().equals("")) {
+            final String newFileName = UUID.randomUUID() + request.image().getOriginalFilename();
+            FileUtil.saveFile(newFileName, request.image());
+            banner.setImage(newFileName);
+        }
         bannerRepository.save(banner);
         log.info("---------------Success Save Banner" + banner + "---------------");
 
     }
 
     @Override
-    public void updateBannerById(Long id, Banner requestBanner, MultipartFile image) {
+    public void updateBannerById(BannerUpdateRequest request) {
         log.info("---------------Save Update---------------");
-        Banner banner = getBannerById(id);
-        banner.setName(requestBanner.getName());
-//        if (FileUtil.saveFile(image.getOriginalFilename(), image))
+        Banner banner = getBannerById(request.id()).setName(request.name());
+        if (nonNull(request.image())
+                && nonNull(request.image().getOriginalFilename())
+                && !request.image().getOriginalFilename().equals("")) {
+            final String newFileName = UUID.randomUUID() + request.image().getOriginalFilename();
+            FileUtil.saveFile(newFileName, request.image());
             FileUtil.deleteFile(banner.getImage());
-            banner.setImage(image.getOriginalFilename());
+            banner.setImage(newFileName);
+        }
         bannerRepository.save(banner);
         log.info("---------------Success Update Banner" + banner + "---------------");
 
@@ -61,9 +86,7 @@ public class BannerServiceImpl implements BannerService {
     public void deleteBanner(Long id) {
         log.info("---------------Save Delete---------------");
         Banner banner = getBannerById(id);
-        if (banner.getImage()!=null){
-            FileUtil.deleteFile(banner.getImage());
-        }
+        if (nonNull(banner.getImage())) FileUtil.deleteFile(banner.getImage());
         bannerRepository.delete(banner);
         log.info("---------------Success Delete Banner" + banner + "---------------");
     }
